Changing reducers setup
