+++ src/reducers/futureStatesReducers.jsx
@@ -1,0 +1,6 @@
+export const futureStatesReducers = (state = {}, action) => {
+    switch(action.type) {
+    default:
+        return state;
+    }
+};
