+++ src/reducers/index.jsx
@@ -1,0 +1,23 @@
+import { combineReducers } from 'redux';
+//import trialTable from 'trialTableReducers';
+//import trialOrder from 'trialOrderReducers';
+//import openTrial from 'openTrialReducers';
+//import timelineOpen from 'timelineOpenReducers';
+//import pastStates from 'pastStatesreducers';
+//import futureStates from 'futureStatesReducers';
+//import dragged from 'draggedReducers';
+import over from './overReducers.jsx';
+
+// Combine all the different reducers into the store
+const rootReducer = combineReducers({
+    //trialTable,
+    //trialOrder,
+    //openTrial,
+    //timelineOpen,
+    //pastStates,
+    //futureStates,
+    dragged,
+    over
+});
+
+export default rootReducer;
