+++ src/reducers/openTrialReducers.jsx
@@ -1,0 +1,2 @@
+export const openTrialReducers = (state = {}, action) => {
+};
