+++ src/reducers/pastStatesReducers.jsx
@@ -1,0 +1,6 @@
+export const pastStatesReducers = (state = {}, action) => {
+    switch(action.type) {
+    default:
+        return state;
+    }
+};
