+++ src/reducers/timelineOpenReducers.jsx
@@ -1,0 +1,6 @@
+export const timelineOpenReducers = (state = {}, action) => {
+    switch(action.type) {
+    default:
+        return state;
+    }
+};
