+++ src/reducers/trialOrderReducers.jsx
@@ -1,0 +1,6 @@
+export const trialOrder = (state = {}, action) => {
+    switch (action.type) {
+    default:
+        return state;
+    }
+};
